[WIP]
This update addresses the tickets #133 and #134.

The mode of the program is now controlled by a new configuration parameter ChatMode, whose value "off" forces the program to use the "Gr." legacy prefix, and "on", which outputs the "chat$group." prefix. For reading the header both prefixes are recognized and accepted.

Function dc_create_outgoing_rfc724_mid has now a new parameter "context" to allow reading the configuration within the function.

The new convention has been reflected on the tests at the end of dc_tools.rs.

In the source the modified code sections are marked by the date and the committer name "Juha-Petri Tyrkkö".